backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
networkx
pytest
pytest-cov
httpx
aiohttp
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]